Add files via upload
Write a class with a main method that uses a priority queue to store a list of chores and the chore’s priority. You need to store the name of the chore and it’s assigned priority. Assigned priorities can be any positive integer. The rest is up to you.